add a 2nd compartment -- ChemotaxisExpressionFlagella -- to chemotaxis_flagella
